+++ DependencyInjection/HttplugExtension.php
@@ -218,6 +218,11 @@ class HttplugExtension extends Extension
         $pluginClientOptions = [];
 
         if ($profiling) {
+            if (!in_array('httplug.plugin.stopwatch', $arguments['plugins'])) {
+                // Add the stopwatch plugin
+                array_unshift($arguments['plugins'], 'httplug.plugin.stopwatch');
+            }
+
             // Tell the plugin journal what plugins we used
             $container
                 ->getDefinition('httplug.collector.plugin_journal')
@@ -340,6 +345,12 @@ class HttplugExtension extends Extension
         $pluginClientOptions = [];
 
         if ($profiling) {
+            // Tell the plugin journal what plugins we used
+            $container
+                ->getDefinition('httplug.collector.plugin_journal')
+                ->addMethodCall('setPlugins', [$name, ['httplug.plugin.stopwatch']])
+            ;
+
             $debugPluginServiceId = $this->registerDebugPlugin($container, $serviceId);
 
             $pluginClientOptions['debug_plugins'] = [new Reference($debugPluginServiceId)];
@@ -348,7 +359,7 @@ class HttplugExtension extends Extension
         $container
             ->register($serviceId, DummyClient::class)
             ->setFactory([PluginClientFactory::class, 'createPluginClient'])
-            ->setArguments([[], $factory, [], $pluginClientOptions])
+            ->setArguments([[new Reference('httplug.plugin.stopwatch')], $factory, [], $pluginClientOptions])
         ;
 
         return $serviceId;